requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
# Flask-OAuthlib==0.9.6
Authlib==1.3.0
bcrypt==4.1.2
pyotp==2.9.0
cryptography==42.0.2
# Werkzeug==2.0.1
python-magic==0.4.27
PyJWT==2.1.0
python-dotenv==1.0.1
python-docx==1.0.1
google-auth-oauthlib==1.0.0
requests==2.31.0
qrcode==7.4.2
Pillow==10.4.0
Werkzeug==3.0.1
okta-jwt-verifier==0.2.3
okta==0.0.4 